Move Loading into shared, fix category bug
