+++ app/actions/resourceActions.js
@@ -8,6 +8,7 @@ import { API_URL } from '../config';
 
 export function fetchResources(categoryId) {
   return function(dispatch) {
+    console.warn("Fetching the resources for " + categoryId);
     dispatch({type: FETCH_RESOURCES});
     fetch(API_URL + "/resources?category_id="+categoryId)
       .then((response) => response.json())
