+++ app/components/shared/index.js
@@ -1,4 +1,5 @@
 // Export the shared components in one place
 export * from './Button';
 export * from './Input';
+export * from './Loading';
 export * from './Map';
